fix(driver): keep pagination counters strongly typed

+++ mobile-driver/lib/main.dart
@@ -981,9 +981,9 @@ class _MesOffresScreenState extends State<MesOffresScreen> with SingleTickerProv
 
   void _changePage(String scope,int delta){
     setState((){
-      if(scope=='active')activePage=(activePage+delta).clamp(0,1<<30);
-      if(scope=='won')wonPage=(wonPage+delta).clamp(0,1<<30);
-      if(scope=='closed')closedPage=(closedPage+delta).clamp(0,1<<30);
+      if(scope=='active')activePage=(activePage+delta).clamp(0,1<<30).toInt();
+      if(scope=='won')wonPage=(wonPage+delta).clamp(0,1<<30).toInt();
+      if(scope=='closed')closedPage=(closedPage+delta).clamp(0,1<<30).toInt();
       _load();
     });
   }
